Update test and remove unreasonable test cases
Currently there are tests to verify the behavior of adding `DateComponents` that sets both `weekOfYear` and `weekOfMonth`. This configuration is ambiguous, and even more so when `wrapping` is `true` since the results vary depending on which component is added first.

Replace tests asserting the behavior of this unreasonable `DateComponents` with those that only uses one kind of `week` component.